Update packages and setup scripts for Python3 and Django 3.2
This updates Django to 3.2.13. In addition, it moves several system
packages into the requirements file and setup scripts and no longer sets
`--system-site-packages`, hence the separate installation of the `wheel`
package.

Note that the GDAL install isn't included in the requirements file
as this must match the version of GDAL on the system, hence the
version check in the setup script; assuming libgdal-dev is installed
and the version matches, the `pip install` should work on recent
versions without setting `C_INCLUDE_PATH` and `CPLUS_INCLUDE_PATH`.

The versions used for the other dependencies are based on the versions
distributed with Debian Bullseye.